requirements.txt
Pillow
requests
scikit-learn
streamlit
numpy
